Update config.py
revert

+++ src/config.py
@@ -1,13 +1,15 @@
 import numpy as np
 
-# Define the paths to all required data files
+# Parameter file path
 PARAMETERS_FILE = "../data/sieve_parameters.txt"
+
+# Data file paths
 DELTA_CURVE_FILE = "../data/delta_curve.npy"
 DYNAMIC_SINE_ENVELOPE_FILE = "../data/dynamic_sine_envelope.npy"
 WITHIN_BAND_MASK_FILE = "../data/within_band_mask.npy"
 ZETA_ZEROS_FILE = "../data/zeta_zeros.npy"
 
-# Define default parameters with corrected Base_Frequency calculation
+# Default parameters for the sieve
 DEFAULT_PARAMETERS = {
     "Amplitude": 12.2,
     "Frequency": 0.001,
@@ -15,5 +17,5 @@ DEFAULT_PARAMETERS = {
     "Tolerance": 0.875,
     "Phase_Shift": 0.0,
     "Seed_Region_End": 50,
-    "Base_Frequency": float(abs(1 / np.log(3) - 1 / np.log(np.pi))),
+    "Base_Frequency": abs(1 / np.log(3) - 1 / np.log(np.pi))
 }
